+++ Source/DesignSystem/index.ts
@@ -12,7 +12,6 @@ export { AlertDialog, AlertDialogProps } from './atoms/AlertDialog';
 export { AlertBox, AlertBoxProps, AlertBoxErrorMessage, AlertBoxInfoMessage } from './atoms/AlertBox';
 export { Button, ButtonProps } from './atoms/Button';
 export * from './atoms/Forms';
-export { DropdownMenu, DropdownMenuProps } from './atoms/DropdownMenu';
 export { Icon, IconProps } from './atoms/Icon';
 export { IconButton, IconButtonProps } from './atoms/IconButton';
 export { Link, LinkProps } from './atoms/Link';
@@ -31,6 +30,7 @@ export { TextField, TextFieldProps } from './atoms/TextField';
 export { EditCell, EditTextFieldCell } from './molecules/DataGrid';
 export { FileUploadForm, FileUploadFormProps, FileUploadFormRef } from './molecules/FileUploadForm';
 export { Graph } from './molecules/Metrics/Graph';
+export { DropdownMenu, DropdownMenuProps } from './molecules/DropdownMenu';
 export { NavigationBar, NavigationBarProps } from './molecules/NavigationBar';
 export { SelectCard, SimpleCard, SimpleCardProps } from './molecules/Card';
 export { SideBar } from './molecules/SideBar';
